número de verificación generado de forma aleatoria

+++ lab02_caceres_fernandez_paslavska/UsuariosRegistrados/AccesoDatos/AccesoDatos.vb
@@ -35,7 +35,7 @@ Public Class AccesoDatos
         BDConexion.Close()
     End Sub
 
-    Public Shared Function IncluirUsuario(pemail As String, nombre As String, apellido As String, preguntaOculta As String, respuesta As String, na As Integer, numeroVerificacion As Integer, verificado As Boolean, codigoGrupo As String, codigoSubGrupo As String, tipoUsuario As String, contraseña As String) As Integer
+    Public Shared Function IncluirUsuario(pemail As String, nombre As String, apellido As String, preguntaOculta As String, respuesta As String, na As Integer, numeroVerificacion As Long, verificado As Boolean, codigoGrupo As String, codigoSubGrupo As String, tipoUsuario As String, contraseña As String) As Integer
         Dim sql As String = "INSERT INTO Erabiltzaileak(email, izena, abizena, galderaEzkutua, erantzuna, na, egiaztatzeZenbakia, egiaztatua, lantaldeKodea, azpitaldeKodea, erabiltzaileMota, pasahitza) VALUES('" & pemail & "', '" & nombre & "', '" & apellido & "', '" & preguntaOculta & "', '" & respuesta & "', '" & na & "', '" & numeroVerificacion & "', '" & verificado & "', '" & codigoGrupo & "', '" & codigoSubGrupo & "', '" & tipoUsuario & "', '" & contraseña & "')"
         SqlComando = New SqlCommand(sql, BDConexion)
         Dim numregs As Integer
